requirements.txt
tiktoken
faiss-cpu
langchain
PyPDF2
python-dotenv
streamlit
openai